Add files via upload
I wrote a blog on Medium ##EDA with Pandas Profiling here is an example with an Automobile data set.

URL	https://medium.com/@david-akinyemi


Pandas profiling is an open-source Python module with which we can easily carry out an Exploratory Data Analysis with few code lines.

It is a great tool to create reports in the interactive HTML format which is quite easy to understand by anyone even if they don’t know programming or data analysis.

In contrast, what pandas profiling does is eliminate the stress of visualizing and understanding the distribution of each variable. It generates a report with all the information easily available.